[@dagster-io/ui] Clean up deps
