+++ js_modules/dagit/packages/core/package.json
@@ -21,10 +21,8 @@
   "peerDependencies": {
     "@apollo/client": "3.3.16",
     "@blueprintjs/core": "^3.45.0",
-    "@blueprintjs/icons": "^3.26.1",
     "@blueprintjs/popover2": "0.10.1",
     "@blueprintjs/select": "^3.16.4",
-    "@blueprintjs/table": "^3.8.33",
     "@dagster-io/ui": "workspace:*",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -57,6 +55,9 @@
     "react-codemirror2": "^7.2.1",
     "react-is": "^17.0.2",
     "react-virtualized": "^9.22.3",
+    "remark": "13.x",
+    "remark-gfm": "1.0.0",
+    "remark-plain-text": "^0.2.0",
     "subscriptions-transport-ws": "^0.9.15",
     "worker-loader": "^3.0.8",
     "yaml": "2.0.0-10"
@@ -74,10 +75,8 @@
     "@babel/preset-typescript": "7.15.0",
     "@babel/runtime": "^7.13.10",
     "@blueprintjs/core": "^3.45.0",
-    "@blueprintjs/icons": "^3.26.1",
     "@blueprintjs/popover2": "0.10.1",
     "@blueprintjs/select": "^3.16.4",
-    "@blueprintjs/table": "^3.8.33",
     "@dagster-io/ui": "1.0.0",
     "@graphql-tools/mock": "^8.5.0",
     "@graphql-tools/schema": "^8.3.1",
