+++ js_modules/dagit/packages/ui/.storybook/blueprint.css
@@ -1,5 +1,3 @@
 @import '@blueprintjs/core/lib/css/blueprint.css';
-@import '@blueprintjs/icons/lib/css/blueprint-icons.css';
 @import '@blueprintjs/select/lib/css/blueprint-select.css';
-@import '@blueprintjs/table/lib/css/table.css';
 @import '@blueprintjs/popover2/lib/css/blueprint-popover2.css';
