+++ js_modules/dagit/packages/ui/package.json
@@ -22,10 +22,8 @@
   },
   "peerDependencies": {
     "@blueprintjs/core": "^3.45.0",
-    "@blueprintjs/icons": "^3.26.1",
     "@blueprintjs/popover2": "0.10.1",
     "@blueprintjs/select": "^3.16.4",
-    "@blueprintjs/table": "^3.8.33",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-is": "^17.0.2",
@@ -34,13 +32,10 @@
   "dependencies": {
     "deepmerge": "^4.2.2",
     "highlight.js": "11.5.0",
-    "lodash": "^4.17.21",
-    "lru-cache": "^7.7.1",
+    "lodash.isequal": "^4.5.0",
     "react-markdown": "6.0.3",
     "react-virtualized": "^9.22.3",
-    "remark": "13.x",
-    "remark-gfm": "1.0.0",
-    "remark-plain-text": "^0.2.0"
+    "remark-gfm": "1.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.16.7",
@@ -72,10 +67,11 @@
     "@types/babel__preset-env": "^7",
     "@types/eslint": "^8",
     "@types/jest": "^27.4.0",
-    "@types/lodash": "^4",
+    "@types/lodash.isequal": "^4",
     "@types/lru-cache": "^7",
     "@types/mdx-js__react": "^1",
     "@types/prettier": "^2",
+    "@types/react-is": "^17",
     "@types/rollup-plugin-node-globals": "^1",
     "@types/styled-components": "^5",
     "@types/testing-library__jest-dom": "^5.14.2",
@@ -92,7 +88,6 @@
     "rollup": "^2.70.1",
     "rollup-plugin-node-globals": "^1.4.0",
     "rollup-plugin-polyfill-node": "^0.8.0",
-    "rollup-plugin-postcss": "^4.0.2",
     "rollup-plugin-styles": "^4.0.0",
     "typescript": "^4.5.4"
   },
